265 feat: Add result error state and setter in ErrorState
The recent code changes added a new boolean state `resultError` and its corresponding setter `setResultError` to the `ErrorState` interface. This state is used to track if there is an error in the result of a query. The changes also updated the `ErrorStore` to include the `resultError` state and setter. Additionally, the `checkGraphQueryInput` hook was modified to return the opposite boolean value for `resultError` based on the value of `plotLeagueQueryStore` and `plotModifierQueryStore`. The `Index` component was updated to use the `resultError` state from the `ErrorStore` and display an error message when it is true. Finally, the `usePostPlottingData` hook was updated to include the `isError` boolean value, which is used to determine if there was an error in fetching the plot data.

This commit message suggests adding the result error state and setter in the `ErrorState` interface, updating the `ErrorStore`, modifying the `checkGraphQueryInput` hook, updating the `Index` component, and updating the `usePostPlottingData` hook.